Add files via upload
Wrapper Code for HTML website to retrieve data via sockets. The Java program acts as a server, and it wraps around the HTML code. Javascript is also used, however, I opted to use the Javascript code only for detecting opened and closed connections. With the change in the status of the connections, the user can be notified via the frontend, the website.